Create functions "upstream_pixel" and "downstream_pixel"
